backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
PyMuPDF
python-docx
python-pptx
google-generativeai
python-dotenv
python-multipart
